extract fetch helpers from client dofetch to shared/utils
