+++ shared/utils/src/index.js
@@ -4,9 +4,12 @@ export * from './clustering.js'
 export * from './common.js'
 export * from './compute.percentile.js'
 export * from './descriptive.stats.js'
+export * from './fetch-helpers.js'
 export * from './fileSize.js'
 export * from './filter.js'
+export * from './hash.js'
 export * from './helpers.js'
+export * from './joinUrl.js'
 export * from './mds.termdb.termvaluesetting' // deprecated?
 export * from './mds3tk.js'
 export * from './roundValue.js'
@@ -22,4 +25,3 @@ export * from './vcf.csq.js'
 export * from './vcf.info.js'
 export * from './vcf.js'
 export * from './vcf.type.js'
-export * from './violin.bins.js'
